Allow deploys of uptous in slack

+++ slack-bot/lib/services/deploy.rb
@@ -120,7 +120,7 @@ module Services
       SH
     end
 
-    def deploy_landing_page
+    def deploy_frekkls
       destination = begin
         if @environment == "production"
           "root@139.59.128.112:/var/www/frekkls.com/html"
@@ -133,6 +133,21 @@ module Services
         cd #{ENV['BUILD_FOLDER']}/core/landing-pages/frekkls && \
         yarn install --silent --no-progress && \
         cp #{ENV['LANDING_PAGE_ENV_FILE']} . && \
+        rm -rf .cache && \
+        node_modules/.bin/gatsby build --no-color && \
+        rsync -azq -e 'ssh -o StrictHostKeyChecking=no -i #{ENV['STATIC_KEY_FILE']}' --delete-after --ignore-errors public/ #{destination}
+      SH
+    end
+
+    def deploy_uptous
+      return "echo fail; false" unless @environment == "production"
+
+      destination = "root@139.59.128.112:/var/www/uptous.co/html"
+      # We should remove the `rm -rf .cache` line when this is fixed: https://github.com/gatsbyjs/gatsby/pull/16220
+      <<~SH
+        cd #{ENV['BUILD_FOLDER']}/core/landing-pages/uptous && \
+        yarn install --silent --no-progress && \
+        cp #{ENV['UPTOUS_ENV_FILE']} . && \
         rm -rf .cache && \
         node_modules/.bin/gatsby build --no-color && \
         rsync -azq -e 'ssh -o StrictHostKeyChecking=no -i #{ENV['STATIC_KEY_FILE']}' --delete-after --ignore-errors public/ #{destination}
